Refactor away from Provider interfaces
This commit refactors away from the IProviderX interfaces and also makes
the fluent interface more logical.

For example, instead of

	Dispatch(new Dispatchable().OnQueue("abc").WithContent("1"))

you now do:

	Dispatch(Dispatchable.OnQueue("abc").WithContent("1"))

and for the Receive method:

	Receive(new Receivable().OnQueue("abc").MessagesToReceive(1))

you now do:

	Receive(1.Message().OnQueue("abc"))

In addition, the fluent interface is also now OPTIONAL. Meaning the
Dispatchable and Receivable objects can be instantiated and passed into
the Dispatch and Receive methods directly.