Refactors and finishes the implementation of the base functionality
This commit refactors a variety of the functionality and also downgrades
the projects from .NET Standard 2.1 to .NET Standard 2.0. This involved
removing the `nullable` configuration option within the project but
allows more people to use the library.

In addition, I finally finished the implementation of the base
functionality. That is, everything works but no providers have been
written.